pyproject.toml
[project]
name = "nlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.0.0",
    "fastapi>=0.116.1",
    "pandas-stubs==2.3.2.250827",
    "scikit-learn>=1.7.2",
    "sentencepiece>=0.2.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision",
    "transformers>=4.56.1",
    "uvicorn>=0.35.0",
]
